just adding one more commit to track progress

+++ src/ofApp.h
@@ -27,7 +27,7 @@ class ofApp : public ofBaseApp{
     
     int boxHeight;
     
-    float nZigZagsY;
+    int nZigZagsY;
     
     int i;
